[refactor/#63] home 컴포넌트 구조 변경
